refactor(acp/tokens): move token management page out from under  and into , this makes it administrator only
