+++ src/controllers/admin/settings.js
@@ -9,8 +9,6 @@ const groups = require('../../groups');
 const languages = require('../../languages');
 const navigationAdmin = require('../../navigation/admin');
 const social = require('../../social');
-const api = require('../../api');
-const pagination = require('../../pagination');
 const helpers = require('../helpers');
 const translator = require('../../translator');
 const plugins = require('../../plugins');
@@ -143,23 +141,6 @@ settingsController.pagination = async (req, res) => {
 settingsController.notifications = async (req, res) => {
 	res.render(`admin/settings/notifications`, {
 		title: `[[admin/menu:settings/notifications]]`,
-	});
-};
-
-settingsController.api = async (req, res) => {
-	const page = parseInt(req.query.page, 10) || 1;
-	const resultsPerPage = 50;
-	const start = Math.max(0, page - 1) * resultsPerPage;
-	const stop = start + resultsPerPage - 1;
-	const [tokens, count] = await Promise.all([
-		api.utils.tokens.list(start, stop),
-		api.utils.tokens.count(),
-	]);
-	const pageCount = Math.ceil(count / resultsPerPage);
-	res.render('admin/settings/api', {
-		title: '[[admin/menu:settings/api]]',
-		tokens,
-		pagination: pagination.create(page, pageCount, req.query),
 	});
 };
