+++ src/views/admin/manage/api.tpl
